mise en place de l'activity component et la refactorisation du code de la page d'accueil

+++ base/views.py
@@ -19,7 +19,10 @@ def home(request):
     )
     room_count = rooms.count()
     topics = Topic.objects.all()
-    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
+    room_messages = Message.objects.filter(
+        Q(room__topic__name__icontains=q)
+    )
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count, 'room_messages': room_messages}
     return render(request, 'base/home.html', context)
 
 
@@ -76,17 +79,17 @@ def logout_user(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    room_messages = room.message_set.all().order_by('created')
+    room_messages = room.message_set.all()
     participants = room.participants.all()
     if request.method == 'POST':
         message = Message.objects.create(
             user=request.user,
             room=room,
             body = request.POST.get('body')
-        )
-
+        ) # create a new message
+        room.participants.add(request.user) #automatically add participant in room
         return redirect('room', pk=room.id)
-    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants }
     return render(request, 'base/room.html', context)
 
 
@@ -123,8 +126,21 @@ def update_room(request, pk):
 @login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Houst batard va !')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     context = {'obj': room}
+    return render(request, 'base/delete.html', context)
+
+@login_required(login_url='login')
+def delete_message(request, pk):
+    messages_del = Message.objects.get(id=pk)
+    if request.user != messages_del.user:
+        return HttpResponse('Houst batard va !')
+    if request.method == 'POST':
+        messages_del.delete()
+        return redirect('room', pk=messages_del.room.id)
+    context = {'obj': messages_del}
     return render(request, 'base/delete.html', context)
